+++ src/layouts/index.js
@@ -5,6 +5,8 @@ import Helmet from "react-helmet";
 import Header from "../components/header";
 import Footer from "../components/footer";
 
+import favicon from "../assets/favicon.ico";
+
 import "./index.scss";
 import "../../pancake/css/pancake.min.css";
 
@@ -13,8 +15,25 @@ const Layout = ({ children, data }) => (
 		<Helmet
 			title={data.site.siteMetadata.title}
 			meta={[
-				{ name: "description", content: "Sample" },
-				{ name: "keywords", content: "sample, something" },
+				{ name: "description", content: "Australian information" },
+				{
+					name: "keywords",
+					content: `australia, ausgov, australia.gov.au, 
+						australian information, australian government`,
+				},
+				{
+					name: "og:title",
+					content: `${data.site.siteMetadata.title}`,
+				},
+				{ name: "og:url", content: "https://ausgov.gov.au" },
+				{ name: "charset", content: "utf-8" },
+			]}
+			link={[
+				{
+					rel: "shortcut icon",
+					type: "image/x-icon",
+					href: `${favicon}`,
+				},
 			]}
 		/>
 		<Header siteTitle={data.site.siteMetadata.title} />